Move RS0019 to its new location.
