+++ src/Roslyn.Diagnostics.Analyzers/Core/Roslyn.Diagnostics.Analyzers.csproj
@@ -114,7 +114,6 @@
     <Compile Include="OverrideObjectEqualsObjectWhenImplementingIequatable.Fixer.cs" />
     <Compile Include="DoNotInvokeDiagnosticDescriptor.cs" />
     <Compile Include="SymbolDeclaredEventMustBeGeneratedForSourceSymbols.cs" />
-    <Compile Include="SymbolDeclaredEventMustBeGeneratedForSourceSymbols.Fixer.cs" />
     <Compile Include="DiagnosticCategory.cs" />
   </ItemGroup>
   <ItemGroup>
